+++ src/Cello/data_State.cpp
@@ -33,7 +33,7 @@ void State::set_cycle(int cycle)
 {
   cycle_ = cycle;
   if (state_type_ == Type::Level) {
-    for (int level = 0; level<cycle_level_.size(); level++) {
+    for (size_t level = 0; level<cycle_level_.size(); level++) {
       set_cycle(cycle,level);
     }
   }
@@ -53,7 +53,7 @@ void State::set_time (double time)
 {
   time_ = time;
   if (state_type_ == Type::Level) {
-    for (int level = 0; level<time_level_curr_.size(); level++) {
+    for (size_t level = 0; level<time_level_curr_.size(); level++) {
       set_time (time,level);
     }
   }
@@ -65,7 +65,7 @@ void State::init_time (double time)
 {
   time_ = time;
   if (state_type_ == Type::Level) {
-    for (int level = 0; level<time_level_curr_.size(); level++) {
+    for (size_t level = 0; level<time_level_curr_.size(); level++) {
       init_time (time,level);
     }
   }
@@ -77,7 +77,7 @@ void State::set_dt (double dt)
 {
   dt_ = dt;
   if (state_type_ == Type::Level) {
-    for (int level = 0; level<dt_level_.size(); level++) {
+    for (size_t level = 0; level<dt_level_.size(); level++) {
       set_dt (dt,level);
     }
   }
